requirements.txt
#python==3.10.0
pyyaml==6.02
plotly==5.24.1
pandas==2.2.3
streamlit==1.41.1